Move bitbucket{,server}, github, gitlab packages into vcs
