--- modules/core/src/main/scala/org/scalasteward/core/bitbucket/http4s/CreatePullRequestRequest.scala
@@ -1,54 +1,0 @@
-/*
- * Copyright 2018-2020 Scala Steward contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.scalasteward.core.bitbucket.http4s
-
-import io.circe.{Encoder, Json}
-import org.scalasteward.core.git.Branch
-import org.scalasteward.core.vcs.data.Repo
-
-private[http4s] case class CreatePullRequestRequest(
-    title: String,
-    sourceBranch: Branch,
-    sourceRepo: Repo,
-    destinationBranch: Branch,
-    description: String
-)
-
-private[http4s] object CreatePullRequestRequest {
-  implicit val encoder: Encoder[CreatePullRequestRequest] = Encoder.instance { d =>
-    Json.obj(
-      ("title", Json.fromString(d.title)),
-      (
-        "source",
-        Json.obj(
-          ("branch", Json.obj(("name", Json.fromString(d.sourceBranch.name)))),
-          (
-            "repository",
-            Json.obj(("full_name", Json.fromString(s"${d.sourceRepo.owner}/${d.sourceRepo.repo}")))
-          )
-        )
-      ),
-      ("description", Json.fromString(d.description)),
-      (
-        "destination",
-        Json.obj(
-          ("branch", Json.obj(("name", Json.fromString(d.destinationBranch.name))))
-        )
-      )
-    )
-  }
-}
