+++ modules/core/src/main/scala/org/scalasteward/core/vcs/VCSSelection.scala
@@ -19,12 +19,12 @@ package org.scalasteward.core.vcs
 import io.chrisdavenport.log4cats.Logger
 import org.scalasteward.core.application.Config
 import org.scalasteward.core.application.SupportedVCS.{Bitbucket, BitbucketServer, GitHub, Gitlab}
-import org.scalasteward.core.bitbucket.http4s.Http4sBitbucketApiAlg
-import org.scalasteward.core.bitbucketserver.http4s.Http4sBitbucketServerApiAlg
-import org.scalasteward.core.github.http4s.Http4sGitHubApiAlg
-import org.scalasteward.core.gitlab.http4s.Http4sGitLabApiAlg
 import org.scalasteward.core.util.{HttpJsonClient, MonadThrow}
+import org.scalasteward.core.vcs.bitbucket.http4s.Http4sBitbucketApiAlg
+import org.scalasteward.core.vcs.bitbucketserver.http4s.Http4sBitbucketServerApiAlg
 import org.scalasteward.core.vcs.data.AuthenticatedUser
+import org.scalasteward.core.vcs.github.http4s.Http4sGitHubApiAlg
+import org.scalasteward.core.vcs.gitlab.http4s.Http4sGitLabApiAlg
 
 final class VCSSelection[F[_]](implicit
     client: HttpJsonClient[F],
@@ -33,13 +33,13 @@ final class VCSSelection[F[_]](implicit
     F: MonadThrow[F]
 ) {
   private def github(config: Config): Http4sGitHubApiAlg[F] = {
-    import org.scalasteward.core.github.http4s.authentication.addCredentials
+    import org.scalasteward.core.vcs.github.http4s.authentication.addCredentials
 
     new Http4sGitHubApiAlg[F](config.vcsApiHost, _ => addCredentials(user))
   }
 
   private def gitlab(config: Config): Http4sGitLabApiAlg[F] = {
-    import org.scalasteward.core.gitlab.http4s.authentication.addCredentials
+    import org.scalasteward.core.vcs.gitlab.http4s.authentication.addCredentials
     new Http4sGitLabApiAlg[F](
       config.vcsApiHost,
       user,
@@ -50,13 +50,13 @@ final class VCSSelection[F[_]](implicit
   }
 
   private def bitbucket(config: Config): Http4sBitbucketApiAlg[F] = {
-    import org.scalasteward.core.bitbucket.http4s.authentication.addCredentials
+    import org.scalasteward.core.vcs.bitbucket.http4s.authentication.addCredentials
 
     new Http4sBitbucketApiAlg(config.vcsApiHost, user, _ => addCredentials(user), config.doNotFork)
   }
 
   private def bitbucketServer(config: Config): Http4sBitbucketServerApiAlg[F] = {
-    import org.scalasteward.core.bitbucket.http4s.authentication.addCredentials
+    import org.scalasteward.core.vcs.bitbucket.http4s.authentication.addCredentials
 
     new Http4sBitbucketServerApiAlg[F](
       config.vcsApiHost,
