+++ modules/core/src/test/scala/org/scalasteward/core/vcs/github/http4s/authenticationTest.scala
@@ -1,0 +1,19 @@
+package org.scalasteward.core.vcs.github.http4s
+
+import cats.Id
+import org.http4s.headers.{Accept, Authorization}
+import org.http4s.{BasicCredentials, Headers, MediaType, Request}
+import org.scalasteward.core.vcs.data.AuthenticatedUser
+import org.scalatest.funsuite.AnyFunSuite
+import org.scalatest.matchers.should.Matchers
+
+class authenticationTest extends AnyFunSuite with Matchers {
+  test("addCredentials") {
+    val request = authentication
+      .addCredentials[Id](AuthenticatedUser("user", "pass"))
+      .apply(Request(headers = Headers.of(Accept(MediaType.text.plain))))
+
+    request.headers shouldBe
+      Headers.of(Accept(MediaType.text.plain), Authorization(BasicCredentials("user", "pass")))
+  }
+}
